go detector
