+++ src/Microsoft.ComponentDetection.Orchestrator/Extensions/ServiceCollectionExtensions.cs
@@ -63,7 +63,6 @@ public static class ServiceCollectionExtensions
         services.AddSingleton<IExperimentConfiguration, SimplePipExperiment>();
         services.AddSingleton<IExperimentConfiguration, RustCliDetectorExperiment>();
         services.AddSingleton<IExperimentConfiguration, VcpkgExperiment>();
-        services.AddSingleton<IExperimentConfiguration, GoDetectorReplaceExperiment>();
         services.AddSingleton<IExperimentConfiguration, PipReportExperiment>();
 
         // Detectors
@@ -81,7 +80,6 @@ public static class ServiceCollectionExtensions
 
         // Go
         services.AddSingleton<IComponentDetector, GoComponentDetector>();
-        services.AddSingleton<IComponentDetector, GoComponentWithReplaceDetector>();
 
         // Gradle
         services.AddSingleton<IComponentDetector, GradleComponentDetector>();
